refactor(): package 구조 정리
