+++ src/main/java/calculator/utils/StringUtils.java
@@ -1,9 +1,5 @@
 package calculator.utils;
 
-
-import org.assertj.core.util.Lists;
-
-import java.io.Serializable;
 import java.util.Arrays;
 import java.util.List;
 import java.util.regex.Matcher;
